Change: copy SSH keys at the end of zapper iot provisioning
