+++ device-connectors/src/testflinger_device_connectors/devices/zapper_kvm/__init__.py
@@ -185,35 +185,6 @@ class DeviceConnector(ZapperConnector):
 
         oemscript.provision()
 
-    def _copy_ssh_id(self):
-        """Copy the ssh id to the device"""
-
-        logger.info("Copying the agent's SSH public key to the DUT.")
-
-        try:
-            test_username = self.job_data.get("test_data", {}).get(
-                "test_username", "ubuntu"
-            )
-            test_password = self.job_data.get("test_data", {}).get(
-                "test_password", "ubuntu"
-            )
-        except AttributeError:
-            test_username = "ubuntu"
-            test_password = "ubuntu"
-
-        cmd = [
-            "sshpass",
-            "-p",
-            test_password,
-            "ssh-copy-id",
-            "-o",
-            "StrictHostKeyChecking=no",
-            "-o",
-            "UserKnownHostsFile=/dev/null",
-            f"{test_username}@{self.config['device_ip']}",
-        ]
-        subprocess.check_output(cmd, stderr=subprocess.STDOUT, timeout=60)
-
     def _change_password(self, username, orig_password):
         """Change password via SSH to the one specified in the job data."""
